HSEARCH-3725 Change GsonProvider so that there's only one implementation
We don't need multiple implementations anymore, and we'll need to
execute complex initialization code to work around a bug in Gson
(https://github.com/google/gson/issues/764).
