+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/gson/spi/GsonProvider.java
@@ -6,21 +6,52 @@
  */
 package org.hibernate.search.backend.elasticsearch.gson.spi;
 
+import java.util.function.Supplier;
+
 import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+
 
 /**
  * Centralizes the configuration of the Gson objects.
- *
  */
-public interface GsonProvider {
+public final class GsonProvider {
 
-	Gson getGson();
+	public static GsonProvider create(Supplier<GsonBuilder> builderBaseSupplier, boolean logPrettyPrinting) {
+		return new GsonProvider( builderBaseSupplier, logPrettyPrinting );
+	}
+
+	private final Gson gson;
+
+	private final Gson gsonNoSerializeNulls;
+
+	private final JsonLogHelper logHelper;
+
+	private GsonProvider(Supplier<GsonBuilder> builderBaseSupplier, boolean logPrettyPrinting) {
+		// Null serialization needs to be enabled to index null fields
+		gson = builderBaseSupplier.get()
+				.serializeNulls()
+				.create();
+
+		gsonNoSerializeNulls = builderBaseSupplier.get()
+				.create();
+
+		logHelper = JsonLogHelper.create( builderBaseSupplier.get(), logPrettyPrinting );
+	}
+
+	public Gson getGson() {
+		return gson;
+	}
 
 	/**
 	 * @return Same as {@link #getGson()}, but with null serialization turned off.
 	 */
-	Gson getGsonNoSerializeNulls();
+	public Gson getGsonNoSerializeNulls() {
+		return gsonNoSerializeNulls;
+	}
 
-	JsonLogHelper getLogHelper();
+	public JsonLogHelper getLogHelper() {
+		return logHelper;
+	}
 
 }
